poprawki

+++ src/Figure/Line2D.java
@@ -8,7 +8,7 @@ public class Line2D extends Shape2D {
         super(p1, p2);
     }
 
-    public Line2D(int x1, int y1, int x2, int y2) {
+    public Line2D(double x1, double y1, double x2, double y2) {
         super(x1, y1, x2, y2);
     }
 }
